First stage of separating PIE-clean into a new repo
