--- pie/exams_loader.py
@@ -1,16 +1,0 @@
-"""
-exams_loader.py
-
-Contains functions for loading and processing exam or assessment data.
-"""
-
-def load_exams_data(data_path: str, source: str):
-    """
-    Load exams data from the specified path. Currently a placeholder.
-
-    :param data_path: Path to exams data.
-    :param source: The data source (e.g., PPMI).
-    :return: A placeholder dictionary or DataFrame representing exams data.
-    """
-    # TODO: Implement actual loading logic.
-    return {"exams_data_placeholder": None}
