--- pie/imaging_loader.py
@@ -1,16 +1,0 @@
-"""
-imaging_loader.py
-
-Contains functions for loading and processing imaging data.
-"""
-
-def load_imaging_data(data_path: str, source: str):
-    """
-    Load imaging data from the specified path. Currently a placeholder.
-
-    :param data_path: Path to imaging data.
-    :param source: The data source (e.g., PPMI).
-    :return: A placeholder dictionary or DataFrame representing imaging data.
-    """
-    # TODO: Implement actual loading logic.
-    return {"imaging_data_placeholder": None}
