--- pie/wearables_loader.py
@@ -1,16 +1,0 @@
-"""
-wearables_loader.py
-
-Contains functions for loading and processing wearable device data.
-"""
-
-def load_wearables_data(data_path: str, source: str):
-    """
-    Load wearables data from the specified path. Currently a placeholder.
-
-    :param data_path: Path to wearables data.
-    :param source: The data source (e.g., PPMI).
-    :return: A placeholder dictionary or DataFrame representing wearables data.
-    """
-    # TODO: Implement actual loading logic.
-    return {"wearables_data_placeholder": None}
