+++ requirements.txt
@@ -1,3 +1,6 @@
+# First stages of data processing
+pie-clean @ git+https://github.com/MJFF-ResearchCommunity/PIE-clean@main
+
 # Core data and ML
 numpy>=1.21.2
 pandas>=1.3.3
